Suppression de commentaires inutiles + correction de fautes d'orthographe + ajout de verification de types (hors String sauf mail) et mise de htmlentity() sur les strings pour tout ce qui peut rentrer dans la bdd

+++ Model/ModelInscriptionEtu.php
@@ -7,26 +7,35 @@ function generationToken(){
 
 
 function ajouterEtudiant($db,$nom, $prenom, $dateDeNaissance, $adresse, $ville, $codePostal, $anneeEtude, $formation, $email,  $ine, $token){
+    $prenom=htmlentities($prenom);
+    $nom=htmlentities($nom);
+    $adresse=htmlentities($adresse);
+    $ville=htmlentities($ville);
+    $formation=htmlentities($formation);
+    $ine=htmlentities($ine);
+    if(filter_var($codePostal,FILTER_VALIDATE_INT) and filter_var($email,FILTER_VALIDATE_EMAIL)){
     $ajout = $db->prepare("INSERT INTO Etudiant (Nom, Prenom, DateDeNaissance, Adresse, Ville, CodePostal, AnneeEtude, Formation, Email, MotDePasse, INE, codemail) 
     VALUES (upper(:nom), :prenom, :dateDeNaissance, :adresse, :ville, :codePostal, :anneeEtude, :formation, :email, null, :ine, :CodeMail)");
-    $ajout->execute(array($nom, $prenom, $dateDeNaissance, $adresse, $ville, $codePostal, $anneeEtude, $formation, $email, $ine, $token));
+    $ajout->execute(array($nom, $prenom, $dateDeNaissance, $adresse, $ville, $codePostal, $anneeEtude, $formation, $email, $ine, $token));}
 }
 
 
 function ajouterCV($db, $etu, $chemin){
-    $sql = "INSERT INTO CV (id, chemin, contenu) VALUES (:id, :chemin, :contenu)";
-    $stmt = $db->prepare($sql);
+    if(filter_var($etu,FILTER_VALIDATE_INT) and file_exists($chemin)) {
+        $sql = "INSERT INTO CV (id, chemin, contenu) VALUES (:id, :chemin, :contenu)";
+        $stmt = $db->prepare($sql);
 
-    if (!empty($etu)) {
-        $stmt->bindParam(':id', $etu);
-        $stmt->bindParam(':chemin', $chemin);
+        if (!empty($etu)) {
+            $stmt->bindParam(':id', $etu);
+            $stmt->bindParam(':chemin', $chemin);
 
-        $stmt->bindValue(':contenu', file_get_contents($chemin), PDO::PARAM_LOB);
-        if ($stmt->execute()) {
-            echo "Le fichier a été ajouté avec succès à la base de données.";
-            header('Location: ../View/ViewEtuMain.php');
-        } else {
-            echo '<div style="color: red;">Erreur lors de l\'ajout du fichier dans la base de données : ' . $stmt->errorInfo()[2] . '</div>';
+            $stmt->bindValue(':contenu', file_get_contents($chemin), PDO::PARAM_LOB);
+            if ($stmt->execute()) {
+                echo "Le fichier a été ajouté avec succès à la base de données.";
+                header('Location: ../View/ViewEtuMain.php');
+            } else {
+                echo '<div style="color: red;">Erreur lors de l\'ajout du fichier dans la base de données : ' . $stmt->errorInfo()[2] . '</div>';
+            }
         }
     }
 }
